Stop one cause of NPCs being stranded in the open
Only throw an exception on trying to process a container twice if we are
in protection, this should stop NPCs being stranded in the open when
hitting small numbers of mines (as well as when the error occurs due to
other causes).
In future it should be fixed such that it distinguishes that something
unusual has happened and handle it properly.