Change port defense calculations.
Upgrade is no longer added after reaching max level (cannot upgrade any
further and do not want the additional defences from it).
Credits in the port now add 500 shields/armour and 50 CDs per 10mil,
this means that a port with loads of money in it will have increased
defences to max and will not lose the defences from excess money when
downgrading (as it would with being over upgraded).