Change exp to be holds / 13 rather than / 15 + 2.
This means that experience gained is much more closely related rather
than the situation we have now with the Stellar Freighter getting the
same experience per trade as the Freighter which has 25 more holds, but
the Leviathan gaining 11% more experience than the Freighter by having
only 5 more holds.
Key comparisons in the old system:
ACV was equal to the ACV.
SF was equal to the Thief.
PT/DS better than the ATM.
Amba much better than the IST.