Link port downgrade more closely to damage done.
You now get more smaller chances of a port downgrading, so doing 1900
damage gives you a better chance than doing 1000 damage (previously they
had the same chance), it means there is a slightly lower chance of
downgrading a port when doing just over an even 1000 damage, however it
is only a small difference.  For 3000 damage previously it was a 38.588%
chance compared to now it is a 36.675% chance, however at 2900 it would
be a higher chance than previously.