trader_relations.php: display all Races
On the "Trader Relations" page, include relations with Neutral, as is
done in the "Relations (Personal)" section of the "Trader Status" page.
There is no reason to make this exception, because relations with
Neutral can be improved (for Alskant) by trading at Neutral ports.
Including this information for non-Alskant traders doesn't hurt
anything, it will just always be zero.