Customize permission to view bonds in the planet list
Add a `view_bonds` field to the `alliance_has_roles` table, which sets
the permission of each alliance role to view the Bonds column of
the Planet List. By default, only the Leader role has this permission
enabled, but this can be modified for each role.

This is an improvement to the previous behavior, where only the player
in the alliance with the "*" could view the bonds, and this was not
customizable.

The Flyway patch increments the database version to V1_6_61_01.