SocialLogin.class.inc: prevent Facebook exception
If not specifying a Facebook App ID, return an empty Facebook
login link. We do this to prevent an exception in the Facebook
authorization code if we were to continue.

Ideally we would perhaps omit the Facebook login object altogether
if there is no App ID, but that would require more extensive changes
to the login page code. For this minimal change, a non-interactive
button is preferable to an Exception.