Fix AJAX clearing military payment message prematurely
Follow the protocol in bounty_claim.php for storing a message from
a one-time transaction (in this case, claiming military payment)
so that it can correctly remain on the screen when AJAX refreshes.

Without this fix, the message about claimed military payment is
immediately replaced with the message that there is no military
payment to claim.