smr.inc: ensure `xmlify` produces valid XML
The htmlentities function is not guaranteed to build legal XML.
We switch to htmlspecialchars with the ENT_XML1 flag.
See https://stackoverflow.com/a/35823299 for more information.

In particular, htmlentities escaped '’' into '&rsquo', which XML
fails to parse with the error: Entity 'rsquo' not defined. This
occurred when preparing the XML for ajax on the Planet Construction
page, which caused the auto-refresh to silently stop. With this
patch, the Planet Construction page now always auto-refreshes.

We only noticed this when trying to build (and so suspected an
issue in the "Currently Building" section), because otherwise the
middle_panel does not change and therefore is not added to the
XML data.