Fix html in planet_construction.php template
The "currently building" section had invalid html: a `<div>` inside
a `<p>`. This caused inconsistent behavior in how it was displayed.

In particular, the "Cancel" button would jump in its position after
the first ajax update due to the difference in how the html is
parsed between a modern browser (e.g. Chrome) and the DOMDocument
class of phpQuery (which supplies the html updates to ajax. To fix
this, we simply remove the `<p>`.

The styling is bad after this change, so we add line breaks to pad
around the Cancel button.