newsletter_send.php: add preview
Provide more details in the admin page about the newsletter that
will be sent (i.e. newsletter number, html text, and plain text).
If there are no newsletters, displays a message saying so, rather
than sending a blank newsletter.

To ensure that the sent newsletter is the same as the previewed
one, we no longer query the database in the processing page;
instead, we _only_ do it in the preview page and then pass that
data onward.

This addresses the preview capabilities requested in issue #39.