Add a salutation to the newsletter message body
Closes #38.

The message body will be prefixed with the salutation and the
player's (real) name when sending to all players. However, when
specifying a single recipient, what you enter will be the full
salutation (since there is no player account to associate it with).

The salutation can be specified (or omitted) on the preview page.