Refactor combat_log_viewer.php
This file had two very distinct modes of operation. One was the
listing of combat logs, and the other was the viewing of the logs.
Breaking this file into two (a "lister" and a "viewer") avoids the
somewhat confusing overloading of `$action` to represent both the
6 menu item pages (the "listers") and `$action==5` (the "viewer").

This is for code clarity only; no functionality changes except
to provide an additional error message when no logs are selected
in the form submission.

Some additional refactoring details of note:

* The new "processing" file handles all the possible ways to
  submit the form ('View', 'Save', and 'Delete'), and redirects
  to the appropriate display page with the $var already set.
  This dramatically simplifies the logic of the display pages,
  which no longer have to check both `$_REQUEST` and `$var` for
  data.

* The "viewer" file no longer has a `$var['direction']` field.
  It is much cleaner to set the `current_log` for the next page
  (when pressing the next/previous log buttons).

* Comments were added liberally to the non-template files.
  Combined with the more digestable file length from breaking up
  one file into three (and a significant reduction in the need for
  nested "ifs"), it should now be much easier to understand
  the logic.