SmrPlanet.class.inc: fix undefined property
Planet armour didn't exist when "delayed" damage was added to planets.
But since "delayed" damage was unused when planet armour was added,
there was no practical impact of the property `delayedArmourDelta`
being undefined. However, since it causes a php warning, we fix
the implementation of delayed armour damage, lest it ever be used.