alliance_pick.php: replace submit form with button
AJAX updates are disabled by Template::checkDisableAJAX for any html
with a `submit` input element. When the form does nothing but create
a link (as it does in this case), then it can be trivially replaced
with a linked button, which does _not_ cause AJAX to be disabled.

Since we conditionally displayed the "Pick" buttons on the draft pick
page, the AJAX updates appeared inconsistent and disrupted the flow of
drafting. This fix allows the page to AJAX update consistently.