Fix bug in saving MICRO_TIME to the database
The `microtime` built-in function outputs a string that looks like:

0.39324900 1504296581

The first word is the number of microseconds in seconds.
The function `microtimeMSec` took the 6 digits after the decimal
from the first word and converted that to an int.

However, if the number of microseconds was small enough to start
with a `0` after the decimal (e.g. `0.09324900`), then casting
the substring to an int would give `93249` instead of `093249`.
Therefore, when reconstructing the time from this value, we would
be one digit off, and you'd get a timestamp from 1974 or so.
[Thanks to Curufir for identifying this issue.]

By having `microtimeMSec` (and `microtimeSec`) return strings instead
of ints, we avoid the casting problem altogether.

We also move the sanity check on the microtime argument format
(i.e. that it has a length of 21 characters) out of `escapeMicrotime`,
which is agnostic to the format, and into `microtime{M,}Sec`, which
extract substrings that depend on the precise string format.

Due to the limited use of `escapeMicrotime`, this bug only affected
saving MICRO_TIME into the `account_has_logs` table. With this fix,
the Log Console now properly displays the event times.