Backwards compatibility for incorrectly stored MICRO_TIME
If `getMicrotime` detects an unexpected format for a microtime in
the database, it will now throw an exception unless `$pad_msec=true`,
in which case it will try to repair the incorrectly stored value
according to the way that `escapeMicrotime` broke it previously
(i.e. adding back 0's that were ignored in a cast to int).

WARNING: if the stored value is the incorrect format for an unrelated
reason, then setting `$pad_msec=true` will report a nonsense microtime
without throwing an exception.

Since the Log Console is sorted by the `microtime` database field,
and since this fix doesn't modify the database directly, incorrectly
stored values will show up in the wrong place (likely the top) of the
table, even though the displayed Time will now be correct.