log_console_detail.php: add 2 more colors
Logs from different accounts are color-coded, but the color cycle
only has 4 colors. To avoid repeated colors if viewing more than
4 accounts, 2 additional colors (light blue and pink) were added.

The colors that are hardest to read on the dark green background
(red and blue) were moved to the end of the color cycle.

NOTE: Another option would be to include a column for the account
name instead of, or in addition to, the colors.