SmrPlanet.class.inc: change inhabitableTime
We don't want planets to start becoming habitable before players start
moving, so we have modified the `inhabitableTime` to be at a random
time after players start accumulating turns, rather than after the
game starts. This is because games often start before turns start
accumulating to give players time to create their traders.

NOTE: we use `max(StartDate, StartTurnsDate)` in case `StartTurnsDate`
can be set to before `StartDate`. This likely is an unnecessary
precaution, but it doesn't hurt anything to be careful here.