universe_create_locations.php: Add category display toggle
In addition to links to jump to specific categories, I was asked
if the categories could be toggled as well. To do this, we give
all `<tr>` of each category a class name and toggle that class
when the heading is clicked.

Use colgroup to ensure the table doesn't change its size as
elements of the table are hidden.