Replace javascript anchor links with pure html
The use of html anchor links is complicated by some custom javascript
that prevents links without a target from being clicked multiple times
(to avoid "back button errors"). However, if we add an explicit target
to the html link (in this case, we add the default: "_self"), then the
javascript selector will skip it and allow clicking on multiple anchor
links without needing to refresh the page.

This recovers the functionality of the javascript anchor links, but
is in pure html, as requested.