Improve collapsible sections
The previous design used `<a href="javascript:;" onclick=...>`,
which had two problems:

1. Because it was an `a[href]` without a target, clicking it
   prevented clicking other links due to some custom javascript.
   (The `href` was only there to get the cursor to change to a
   pointer so that it was clear it was clickable.)

2. You could only click on the header text of the first column.

To fix both of these issues, we change the entire header row into
a clickable element (so it's no longer an `a[href]`). Since `<tr>`
can't have an `href` to trick it into becoming a pointer (this is
a bad practice anyway), we give it a `collapsible` class and set
the hover property to change the cursor to a pointer.