More graceful handling of multi-category locations
When locations were split into categories in galaxy creation, we had
removed the ability to list all the details of a location that is
a member of multiple categories.

The categorization is too useful to forego, so we simply place the
location in the first category it matches, but now also add a note
to the location if it matches other categories.

This is not ideal, because when you are looking for Banks, for example,
you do not expect to need to look in Hardware (though an admin could
certainly make such a change to a Bank in 'Edit Locations').

If multi-category locations ever become the norm, then this logic
should be reworked. However, for now, while multi-category locations
do not exist, it is perfectly acceptable.