message_view.php: do not bbify player messages
Messages sent by game services (i.e. planet, scout, etc.) can still
be processed by `bbifyMessage`, but messages sent by players must not.

Now that we have a bb-tag for more sensitive information (such as
combat logs), if player messages are bbify'd, then you could send a
message to another player with "[combatlog=X]", where X is some valid
combat log ID, and the recipient would then have a link to that combat
log -- even if they would otherwise be unable to see it in the combat
log viewer.

This exploit is unlikely to be discovered (and even harder to abuse),
but since it is theoretically possible, it needs to be prevented.

A more comprehensive solution to the exposed combat log problem is
to have a layer between the combat log link and its page that verifies
that the player should be able to access it. Once that is implemented,
it will be simple to allow BBLinks in player messages again.