Kill/Death messages are now sent by a Federal Clerk
Add `sendMessageFromFedClerk` to send from a Federal Clerk instead
of the killing/dying player.

It makes more sense for a third-party agency to send these messages.
(As a pleasant side effect, the messages will no longer show up
in the players' Sent box.)

We also use it to replace `sendMessageFromPort` for consistency,
but we retain this function in case we want ports to send messages
in the future.

Since these messages are now distinct in the code from standard
player messages, we can safely extend their lifetime to a year
(to match admin messages) so that players can keep their kill/death
messages all game. This was a requested feature.