Kill/Death messages are now sent by a Federal Clerk
Add `sendMessageFromFedClerk` to send from a Federal Clerk instead
of the killing/dying player.

This change is needed so that the messages can contain BBLinks
(which will be disabled for player messages).

An interesting side effect is that the kill/death messages will
no longer show up in the player's Sent messages, which didn't
really make sense anyway.