message_view.php: omit notify button if no sender
In addition to omitting the "Reply" and "Blacklist Player" buttons
if there is no sender, omit also the notify button. This should
not be necessary for procedurally generated messages.

The only possible need would be for reporting bugs, but the
intended use of the notify button is for inappropriate player
messages. Bugs can be reported more easily in other ways.