Security layer for combat log BBCode links
Now that we have a BB-tag that links to sensitive information (in this
case combat logs), since player messages are bbify'd, you could send
a message to another player with "[combatlog=X]", where X is some valid
combat log ID, and the recipient would then have a link to that combat
log -- even if they would otherwise be unable to see it in the combat
log viewer.

This exploit is unlikely to be discovered (and even harder to abuse),
but since it is theoretically possible, it needs to be prevented.

To do this, we create a security layer between the BBLink and the
displayed combat log page. This layer verifies that the player should
be allowed to access it. Currently, restrictions include:

* Log is from the current game
* Attacker/Defender is the player or in the player's alliance