galactic_post_view_article.php: fix news post
Prevent spamming Breaking News with Galactic Post articles.

Clicking "Add to Breaking News" would set `$var['news']`, and then
forward that to the current page. Whenever this variable is set,
it adds a news entry to the database. With AJAX, this variable
remains set as the page reloads, and so it adds to the database
over and over again.

To fix this, we simply unset `$var['news']` after the database
is updated.