Add a non-selectable default status in the dropdown
It was possible to accidentally set the status for feature requests
because it was set to "Implemented" by default. By adding a non-
selectable default with an empty value, we ensure that setting
the status only occurs intentionally.

In the processing, we now throw an error if "Set Status" was
clicked, but no status was chosen.