alliance_message.php: fix bank report author
There was an undefined variable `sender_id` for alliance Message
Boards with a Bank Report thread present. This was because the
`sender_id` for bank reports was changed to `ACCOUNT_ID_BANK_REPORTER`,
but the message board code was still using old (numeric) account ID's
to determine the author.

Switching to the newer named constant account ID's fixes the
php warning, but also now correctly identifies the author of Bank
Reports as "Bank Reporter".