bank_anon_processing.php: stay on account page
It is clear from the existing code that it was intended for the
processing page to return to the account transaction page.
However, due to a misspelled variable name, it logged you out
of the account and returned you to the anon account landing page.

By fixing the variable name, a withdraw/deposit no longer kicks
you out of the account transaction page.

The password is not needed or used on the processing page, so
we no longer pass it in the processing container.