Store a Discord User ID in SmrAccount
For now, we restrict to a 1-to-1 mapping between accounts and
Discord ID's. We will handle the case of multi support later.

* SmrAccount: add `{get,set}DiscordId` methods
* MySQL database: add `discord_id` field to `account` table
* Preferences: add field to change/delete Discord ID

We add the Discord ID to the database now in preparation for
forthcoming Discord integration. Having the database updated now
will make it easier to test the new infrastructure.

This change requires flyway migration.