log_console.php: fix PHP warning
The `$var['account_ids']` index is only set when clicking the
back link from `log_console_detail.php`. So when first entering
the page, you get an undefined index error.

Since it is either undefined or an array, there is no point in
checking `is_array`, when `isset` accomplishes the same goal
and avoids the warning. (If for some reason code is changed later
and it is set but not an array, we'll get a warning.)