Add Gruntfile.js
This specifies three tasks to be performed:

* uglify - minifies JavaScript
* cssmin - minifies CSS
* cacheBust - replaces js/css files with hashed filenames

These operations will allow us to ensure that we are serving
optimized js/css files and that we only re-serve them when
their contents have changed.

The tasks are set up in such a way that if they are skipped,
then the js/css links in the source code will still point to
the correct (unminified) files. This is necessary to support
dev environments where the source code is mounted rather than
built.