SmrPlanet.class.inc: change getCompletionModifier formula
Make the `getCompletionModifier` formula less backloaded. The
early builds are increased by a constant 1% term, and the later
builds are decreased by changing the polynomial term from cubic
to parabolic.

To keep the total exp earned by building a planet to max roughly
constant, we need to bump up the base `exp_gain` a little.