course_plot_result.php: forward to current sector
If the start sector is not the current sector or if you are on
a planet, then the existing "Plot A Course" results page is
displayed.

Otherwise, you can immediately navigate the plotted course, so go
directly to the current sector page (which includes the plotted
course summary and action buttons) instead of displaying the
intermediate results page.

Add a "Plotted Course" heading to the course results minipage
on the current sector page. This will help identify the course
results to anyone who may never have seen the intermediate
course results page before.

This change helps to streamline the flow of the game.