channel_msg_money.php: prohibit allianceless use
Because we use `alliance=0` for allianceless players, this
function reports on ALL allianceless players if the command
is run by someone who is allianceless.

To fix this, we simply demand that it only be usable by
players in an alliance.

An allianceless version of this function would make sense
(especially for the planet/bond reporting), but is not
a high priority.