Enable Planet Main hotkey while on planet
To do this, we need to call `Globals::getCurrentSectorHREF`,
because this is the function that adds the current sector hotkey,
and current_sector.php now forwards to planet_main.php if we're
on a planet.

We don't want to create a new hotkey for Planet Main because
we want to use the same key as for current sector.