Add chat_sharing.php
Provides a new database table `account_shares_info` and an interface
to manage this table.

This will be used to allow users in chat to display information
about more than just the account associated with their Discord ID
(or IRC Nick). It is generic, but the primary motivation is to
allow players to display their main and multi.