changelog_view.php: move continue button to top
Players that haven't logged in for a long time are confronted
with a (potentially) massive list of changes. When the continue
button is at the bottom of the page, this creates a pretty large
barrier of entry.

The continue button at the top of the page ensures that people
a) will immediately know that they have the option to continue
and b) aren't forced to scroll through a list of changes if
they don't care about them.

If the list of changes is longer than 420px (just before it
extends beyond the normal height of the game table), the change
log will turn into a scrolling element.