trader_status: change Bounties display
It was not clear whether the bounties listed on the Trader Status
page were bounties on your head or bounties you could claim.

It was also strange that the information on the Trader Status page
and on the View Bounties page (navigable by clicking on the
"Bounties" heading) displayed disjoint information. One showed
bounties on you, the other bounties you could claim.

We restructure the display in the following way:

* Provide information about both bounties on you and bounties you
  can claim on _both_ pages.

* Information is summarized on the "Trader Status" page, and then
  given in complete detail in the "View Bounties" page.