preferences_processing.php: allow name case change
When altering your player name, a check is done to make sure that
the new name is not already in use. Since `player_name` has a case-
insensitive collation, this prevents players from changing the case
of their name.

With this change, we allow the old and new names to be the case-
insensitively the same so long as the account ID's match.