Dumping cargo at 0 exp damages armour
* When at 0 experience, dumping goods does armour damage, but
  will not let you dump if it would take you below 1 armour
  (i.e. you can't destroy your ship dumping cargo).

* Armour damage will be holds/5 (e.g. you can dump 4x 300 holds
  in an IST before you need to UNO). This number may need to be
  adjusted based on feedback.

* There is no rollover. You either lose exp (down to 0) or armour.

* When losing experience, don't show that you lost more experience
  than you have.

The official motivation for this change is to ensure that there
is a tradeoff when slaving. When the slave has exp, it can be lost;
but when you slave at 0 exp, there is no longer any tradeoff.
You basically get slaving for free, which was not intended.
Incurring a ship damage penalty seems like a fitting solution.

More importantly, this idea (suggested by Page) has too perfect of
a role-playing payoff to pass up.