Add button to calculate jump drive turns
A new option for jump drives is "Calculate Turn Cost". When clicked
it will take you to a new page that says how many turns will be
spent when jumping to the target sector.

I think this is a reasonable addition because previously jump drives
had a fixed turn cost (15), so we could report the cost exactly.
Now the cost is just listed as (10+), which is a little misleading
since the turn cost has no upper bound.