Add Directory of Games voting site
Generalize voting infrastructure to arbitrary external URL's
and account for multiple methods of using the callback script.

Add a new star.png, which is the template we will use to create
any new `*_vote.png` images for the voting sites. This is needed
because we don't have the star image that was used to create the
originals.