Use production php.ini by default
By default, use the php.ini-production configuration that comes
with the PHP source code. It should be tuned for performance
compared to the stock `php-apache` image options. This is the
php.ini file used by the current (non-docker) live server.

You can omit this step in the build process by specifying the
build argument `PHP_DEBUG=1`, which can be done in the `.env`
file if you are using the docker-compose setup.

Note that if you use php.ini-production, then you will not
have PHP warnings/errors displayed in the browser even if
you set the SMR configuration value `ENABLE_DEBUG=true` in
the `config.specific.php` file.