Use list.js to sort personal and alliance force lists
This dramatically improves performance of sorting the force lists,
which gets quite slow once the list is large (as it always is).

We also convert forces_list.php and alliance_forces.php to templates
in the process.

NOTE: in doing this, I believe we lose the ability to secondary sort.
While this was useful, the page loading is so slow that it loses its
convenience (you need to load the page 4 times just to sort by
desc mines with asc expire time, which is typically the only case
where secondary sorting is used). However, secondary sorting can
be re-implemented by adjusting the list.js sort function, if it is
requested.