Do not require a composer install for dev env
In the dev environment (where the local source code is mounted in
the container), we previously mounted the repository root.

This had the unpleasant side effect that you had to run `composer
install` in your host environment, which requires you to download
and correctly configure all the PHP dependencies, defeating the
purpose of using docker in the first place.

However, if we mount only the relevant source code directories,
then you retain the `vendor` directory that is created by the
`composer install` in the image, and then you no longer need to
install _any_ dependencies on the host. This is well worth the
slightly enlarged list of mounts.