ajax.js: fix refresh ignoring disable
The issue
=========
When "Auto Refresh" is disabled in preferences, then the
`startRefresh` function will not be called. However, when the
browser regains focus, it will begin to refresh with NO delay.
This is a result of the way the `onfocus` and `startRefresh`
functions interact in ajax.js.

The solution
============
One way to solve this problem is to disable the refresh by
default in ajax.js, and ONLY enable it if `startRefresh` is
called. This ensures that ajax updates will only occur if the
"Auto Refresh" preference is enabled.