Change visibility of (Trade,Goods) to HOF_ALLIANCE
Players can track what ship you're in with public visibility of
number of goods traded, which is not an intended side effect of
the Current HoF.

This HoF category was requested to be removed, but that was likely
before HOF_ALLIANCE visibility was implemented. The visibility change
should be a sufficient solution.