message_view.php: fix ordering of scout messages
The template was previously set up to _always_ display individual
messages before grouped messages, which meant that if the only
new (*) messages were grouped, they would display at the end of
the list, even if it meant that old individual messages displayed
above them. This is not correct, because new messages should
always display at the top.

The engine starts to set this up correctly, by adding new messages
to the message array first, then unread messages second. However,
it then puts them into separate arrays ('Messages' and 'Grouped
Messages'). If we instead put them into the same array, then
the ordering works out correctly in a completely natural way.

As a result of this simple change, we can actually remove _all_
special casing for grouped messages in the template.