Switch to offical Flyway docker image
An officially docker image for Flyway recently became available.
We will use the alpine version, since it is significantly smaller
(100MB from ~600MB).

Note that the latest version is Flyway 5, so if you used an earlier
version on an existing database, you will get a warning about the
Flyway table name change (which will be an error in Flyway 6).

While the warning recommends changing the `flyway.table` property,
it is simpler to just rename the table from `schema_version` to
`flyway_schema_history`.

Flyway 5 finally provides a command to undo migrations. Yay!