Fix mysqli + compatibility databases
One of the issues with mysqli was that I removed the db selection
in the MySqlDatabase constructor, because I noticed that mysqli was
able to do the db selection within the `connect` method.

However, what I didn't realize was that db selection line was used
to switch between databases in compatibility mode with prior game
version databases. Naturally I didn't have these historical
databases in my dev branch, and so the issue never came up for me.

That part of the code is now re-instated and freshly documented.

This fixes d837862aff.