planet_ownership.php: add message about losing ownership
When you go to claim a planet, it is often annoying to check to
see if you already own a planet (especially now that we have multis).

With this change, a message will be displayed letting you know that
if you take ownership when you already have a planet, you will lose
ownership of that other planet. This implicitly declares the
"one planet per person" rule, which was otherwise not stated anywhere.

Also minor rewording to make the new text fit better in context.