Use getAllianceName(true) where possible
The first argument to `Smr{Player,Alliance}::getAllianceName` is
a bool which, if true, turns the name into a link to the Roster.

Using this option simplifies a few places in the code.

NOTE: This replaces "(none)" in a few spots with "No Alliance",
in particular the Current Players and Trader Search Results pages.
The disparity was a result of not everything using the standard
methods. I somewhat prefer the former, but may wait for feedback
before I change it.