Change port raid properties
This change is calibrated to keep the following the same:
 * Level 9 ports (average damage, total defense, loot)
 * Profit per turn for all port levels (as per Page's diagnostics)

While keeping those things fixed, we want to make low level
ports a bit easier. To achieve this, we make the following
tradeoffs for low level ports:
 * Decreased port defenses (fewer base CD's)
    Level 1: 1000 -> 500
 * Slightly increased port offense (higher avg. turret damage)
    Level 1: 150 -> 450
 * Very slightly decreased loot
    Level 1: 4.8 mil -> 4.0 mil
 * Significantly decreased max turret damage
    Level 1: 6 * 250 damage -> 4 * 125 damage

Decreased max damage for lower level ports is important to make
them much less risky (since, e.g., losing a ship is never worth
the ~$4 mil you loot from the port). To achieve this, we make
the following changes to port turrets:
 * All turrets have the same accuracy (10%*Level -> 90%)
 * Number of turrets is now variable (6 -> 3+Level)
 * Turret damage is halved (250 -> 125)

For full details, see:
https://docs.google.com/spreadsheets/d/1LAt0H8G8IHeGhwIIRtpf8gU1_mo3ardzM19kPZ5-4hQ

Cosmetic change:
 * Rename "Port Turret" to "Anti-Theft Sentry Turret"
 * Rename "Planet Turret" to "Zenith Turret" (while we're there)
 * Add a restriction to these weapons so that they can be
   displayed in the weapon list as being restricted to ports/planets.