port_attack_warning.php: fix shifting elements
This page contained a `<table>` element inside a `<p>`, which is
invalid HTML. There is no standard way to parse this, and the
DOMDocument parses it differently than the browser. This
results in the `<p>` element shifting after the auto-refresh.

We replace the `<p>` with `<br />` to avoid this problem.