phplint.sh: correctly check for errors
The `error` variable was being set to `true` in a subshell,
since it was done after a pipe. This resulted in `error=false`
in the main shell no matter what. We restructure the script
to ensure that the do-while loop is run in the main shell.

Also change the error flag variable, to make it clearer that
there is no boolean anything here. It's string comparison.