SmrPort: restructure to permit serialization
Closes #281.

Serialize without modifying the port.

Break the `goods` member into three more logically distinct
members: `goodIDs`, `goodAmounts`, and `goodDistances`.
The `goodIDs` are broken up into "Buy", "Sell", and "All Goods"
arrays as before, but the latter two are simple assoc. arrays.

We will omit `good{Amounts,Distances}` in the serialization,
and only keep the `goodIDs` (which is all the cached port
users need).

We can also avoid modifying `cachedVersion` by setting this
to true in `__wakeup` instead of `__sleep`. This is because
the port we are serializing is not cached -- only the one
we get when we unserialize.

Code is included in `__wakeup` to temporarily permit unserializing
old port caches (this can be removed after any running game is
over).

(Note: below, a "good" means the lengthy data array that
`Globals::getGood{,s}` returns.)
Due to the modified internal data structure of `SmrPort`, some
of the methods have been changed as well:
* `restockGood` takes a goodID instead of a "good".
* `getLootGoodHREF` now takes a goodID instead of a "good".
* `getGoods` is now private, and is used only to populate
  the `good{IDs,Amounts}` members in the ctor.
  It also throws an exception if called on a cached port
  instead of returning false.
* `getVisibleGoods` is now private, and is just a common
  function for `getVisibleGoods{Sold,Bought}`.
* `getVisibleGoodsAll` has been removed.
* `getVisibleGoods{Bought,Sold}` now return an array of goodIDs
  instead of an array of "goods".
* `getGoods{All,Sold,Bought}` have been replaced with
  `get{All,Sold,Bought}GoodIDs`, and they return an array of goodIDs
  instead of an array of "goods".
* `getGoodStatus` has been renamed `getGoodTransaction` to more
  clearly describe what it does.
* `addCachePorts` no longer needs to store and then reset the
  `$cachedVersion` and `$goods`.
* `__sleep` no longer modifies the instance members, and replaces
  `goods` and `cachedVersion` with `goodIDs` in the serialized
  members.
* `__wakeup` now sets `cachedVersion=true` instead of `__sleep`.

Some unnecessary mysql queries have also been removed from `SmrPort`:
* updating `port_has_goods` amounts in `restockGood` (this is
  redundant with `update` called by `savePorts`?)
* updating `port_has_goods` amounts in `setGoodAmount`.