SmrAccount: make is{Newbie,Veteran} less expensive
The previous implementation of these methods used `getRank`,
which has to query the user stats in the database. This gets
especially expensive when you're doing it for a list of players:

* Current Players list
* Rankings
* Alliance rosters
* Current Sector Players list

By switching to `maxRankAchieved` for these methods, we can
use data that we already have for the `SmrAccount` (no extra
database calls), which significantly speeds up these pages.

This does mean, however, that there may be a time when a player
has technically reached Fledgeling (the first veteran rank) but
is still reported in italics (Newbie/Beginner). I do not think we
need to be rigorous about this, because it is never reported
inconsistently and the player's `maxRankAchieved` will be updated
next time they log in (or view Rankings page, or Trader Status page).