Globals: restructure isValidGame
The previous call chain for many in-game functions was:

getGameType -> isValidGame -> getGameInfo

However, `getGameInfo` was a fairly expensive function, constructing
a large array for every call, only to have `isValidGame` check if
the result is empty or not. The condition for whether or not it was
empty was if the `SmrGame` constructor threw an exception.

To speed up all these calls, we instead do the `SmrGame` constructor
check in `isValidGame`, so `getGameType` calls no longer need to
do the work from `getGameInfo`.

We also simplify the `SmrGame` ctor check by adding the
`GameNotFoundException` class, and throw that instead of the base
`Exception` if the game is not found.